+++ CHANGES.md
@@ -11,6 +11,7 @@ Changes in 7.0
 - CSS fix align "+" New Event icon to bottom in Calendar.php, CalendarDay.inc.php, stylesheet.css & zreponsive.css
 - Fix Day Number when multiple calendars and school years in CalendarDay.inc.php, DayToNumber.inc.php
 - Fix Numbered days display in SchoolPeriodsSelectInput.fnc.php & Courses.fnc.php
+- SQL improve Numbered days in AddAbsences.php, Administration.php, DailySummary.php, TakeAttendance.php, TeacherCompletion.php & UpdateAttendanceDaily.fnc.php
 - Place Rollover under Utilities separator in Menu.php
 - Merge Schedule Report & Master Schedule Report in Menu.php, MasterScheduleReport.php, ScheduleReport.php & rosariosis.sql
 - Add Students column to report in RequestsReport.php